requirements.txt
numpy==1.23.5
pandas==1.5.3
tqdm==4.64.1
requests==2.28.2
Pillow==9.3.0
pytesseract==0.3.10
torch>=1.12.0
transformers==4.24.0
tensorflow==2.11.0
keras==2.11.0
scikit-learn==1.2.2